test-requirements.txt
nose==1.3.7
requests==2.18.4
PyHamcrest==1.9.0
